pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "reconhecimento-facial"
version = "0.1.0"
description = "Detecção facial com modelos locais e geração de legendas"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [{name="Projeto OpenAI"}]

dependencies = [
    "opencv-python",
    "mediapipe",
    "ultralytics",
    "torch>=2.6",
    "torchvision",
    "facenet_pytorch",
    "transformers",
    "huggingface_hub",
    "python-dotenv",
    "numpy",
    "flask",
    "questionary",
    "face_recognition; platform_system != 'Windows'",  # optional
    "psycopg2-binary",
    "openai-whisper",
    "sounddevice",
]

[project.scripts]
face-detect = "reconhecimento_facial.face_detection:main"
face-app = "reconhecimento_facial.app:main"
whisper-translate = "reconhecimento_facial.whisper_translation:main"

[tool.pytest.ini_options]
addopts = "-vv"
